% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome.R
\name{chromosome}
\alias{chromosome}
\title{Extracts the chromosome numbers}
\usage{
chromosome(genabel.data)
}
\arguments{
\item{genabel.data}{A GenABEL-like data of class gwaa.data2.}
}
\value{
Returns an array of chromosome numbers
}
\description{
Gets the chromosome numbers.
}
\author{
Lars Ronnegard
}
