% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_GRM2.R
\name{compute.GRM}
\alias{compute.GRM}
\title{Computes a Genetic Relationship Matrix from a GenABEL-like object}
\usage{
compute.GRM(gen.data)
}
\arguments{
\item{gen.data}{The GenABEL-like object.}
}
\value{
Returns a genomic relationship matrix.
}
\description{
One method for GRM computations implemented.
}
\author{
Lars Ronnegard
}
