% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_gwaa_data2.R
\name{Create_gwaa_data2}
\alias{Create_gwaa_data2}
\title{Creates a gwaa.data2 object}
\usage{
Create_gwaa_data2(genotypes, chromosome = NULL, map = NULL, phenotypes = NULL)
}
\arguments{
\item{genotypes}{A matrix with genotype values coded as (0,1,2) or (-1,0,1)}

\item{chromosome}{An array of characters for the chromosomes. Length equal to the number of SNPs.}

\item{map}{An array with the order of the SNPs.}

\item{phenotypes}{A data frame including columns with phenotypes and a column with ids, called "id"}
}
\value{
Returns a gwaa.data2-object.
}
\description{
Creates a gwaa.data2 object from input.
}
\author{
Lars Ronnegard
}
