% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Create_gwaa_scan.R
\name{Create_gwaa_scan}
\alias{Create_gwaa_scan}
\title{Creates a scan.gwaa object}
\usage{
Create_gwaa_scan(data, P1df, SNP.eff)
}
\arguments{
\item{data}{A gwaa.data object}

\item{P1df}{P-values computed from external analysis}

\item{SNP.eff}{Estimated additive SNP effects}
}
\description{
Creates a scan.gwaa object from a GenABEL object and p-values.
}
\author{
Lars Ronnegard
}

