% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_gwaa_scan2.R
\name{Create_gwaa_scan2}
\alias{Create_gwaa_scan2}
\title{Creates a scan.gwaa2 object}
\usage{
Create_gwaa_scan2(data, P1df, SNP.eff)
}
\arguments{
\item{data}{A gwaa.data2 object}

\item{P1df}{P-values computed from external analysis}

\item{SNP.eff}{Estimated additive SNP effects}
}
\value{
Returns a scan.gwaa2 object.
}
\description{
Creates a scan.gwaa2 object from the rGLS output.
}
\author{
Lars Ronnegard
}
