% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estlambda2.R
\name{estlambda}
\alias{estlambda}
\title{Function to estimate lambda}
\usage{
estlambda(data, plot = FALSE, method = "regression", filter = TRUE)
}
\arguments{
\item{data}{An array of P-values}

\item{plot}{Logical. TRUE to produce a plot}

\item{method}{Either "regression" or "median".}

\item{filter}{Logical. If TRUE the extreme P-values are not included in the estimate of lambda.}
}
\value{
Returns a list with estimate and standard error.
}
\description{
Estimates lambda from P-values. Most code copied from the archived GenABEL package
}
\author{
Lars Ronnegard
}
