\name{flycatchers}
\alias{flycatchers}
\docType{data}
\title{
10,000 SNPs from 849 collared flycatchers
}
\description{
A GenABEL object of class gwaa.data. Only ID and sex included in phdata(flycatchers).
}
\usage{data(flycatchers)}
\examples{
data(flycatchers)
summary(flycatchers)
}
\keyword{datasets}
