% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwaa.data2.R
\docType{class}
\name{gwaa.data2-class}
\alias{gwaa.data2-class}
\title{An S4 class to represent GWAS input data}
\description{
An S4 class to represent GWAS input data
}
\section{Slots}{

\describe{
\item{\code{phdata}}{Phenotype information including id}

\item{\code{gtdata}}{object of class \code{\link{snp.data}} with genotype information}
}}

