% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idnames.R
\name{idnames}
\alias{idnames}
\title{Extracts the id names}
\usage{
idnames(genabel.data)
}
\arguments{
\item{genabel.data}{A GenABEL-like data of class gwaa.data2.}
}
\value{
Returns an array with the names of the individuals (as character).
}
\description{
Gets the idnames.
}
\author{
Lars Ronnegard
}
