% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_gwaa_data.R
\name{keep_gwaa_data}
\alias{keep_gwaa_data}
\title{A function to subset an gwaa.data2 object}
\usage{
keep_gwaa_data(genabel.data, indx.keep = NULL)
}
\arguments{
\item{genabel.data}{A GenABEL-like data of class gwaa.data2.}

\item{indx.keep}{Indeces to extract.}
}
\description{
Extracts a subset of the data.
}
\author{
Lars Ronnegard
}
