% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Extracts the map information}
\usage{
map(genabel.data)
}
\arguments{
\item{genabel.data}{A GenABEL-like data of class gwaa.data2.}
}
\description{
Gets the map.
}
\author{
Lars Ronnegard
}
