% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nids.R
\name{nids}
\alias{nids}
\title{Extracts the number of ids}
\usage{
nids(genabel.data)
}
\arguments{
\item{genabel.data}{A GenABEL-like data of class gwaa.data2.}
}
\value{
Returns the number of individuals.
}
\description{
Gets nids.
}
\author{
Lars Ronnegard
}
