% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scan.gwaa2.R
\name{plot.scan.gwaa2}
\alias{plot.scan.gwaa2}
\title{Function to plot P-values as a Manhattan plot}
\usage{
\method{plot}{scan.gwaa2}(
  x,
  y,
  ...,
  ystart = 0,
  col = c("blue", "green"),
  sort = TRUE,
  ylim,
  main = NULL
)
}
\arguments{
\item{x}{A scan.gwaa2 object created by the rGLS function}

\item{y}{A parameter not used in the current version}

\item{...}{Possible additional parameters (not used in the current version)}

\item{ystart}{Lowest value on the y-axis}

\item{col}{Default is c("blue","green")}

\item{sort}{Logical. If TRUE the SNPs are sorted before plotting.}

\item{ylim}{Limits of the y-axis}

\item{main}{Plot title}
}
\value{
No return value, called for side effects
}
\description{
Creates a Manhattan plot
}
\author{
Lars Ronnegard
}
