% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.gwaa2.R
\docType{class}
\name{scan.gwaa2-class}
\alias{scan.gwaa2-class}
\title{An S4 class to represent SNP data}
\description{
An S4 class to represent SNP data
}
\section{Slots}{

\describe{
\item{\code{results}}{The results from the rGLS function as a data.frame}

\item{\code{lambda}}{Computed inflation factor as list}

\item{\code{idnames}}{Idnames as character}

\item{\code{map}}{SNP order as numeric}

\item{\code{chromosome}}{The chromosome name for each SNP as numeric}

\item{\code{call}}{The call made by rGLS as call}

\item{\code{family}}{The assumed distribution of the outcome. Only "gaussian" allowed.}
}}

