% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gendata.R
\name{simulate_gendata}
\alias{simulate_gendata}
\title{Function to simulate genotype data for the RepeatABEL package.}
\usage{
simulate_gendata(n = 100, p = 1000, LD = 0.9, n.chrom = 1)
}
\arguments{
\item{n}{Number of individuals.}

\item{p}{Number of SNPs.}

\item{LD}{An LD-parameter. LD=1 gives complete LD and LD=0 no LD.}

\item{n.chrom}{The size of a simulated SNP.effect.}
}
\value{
Returns a gwaa.data2 object.
}
\description{
The function simulates n individuals and p SNPs, with linkage disequilibrium (LD) given by the LD-parameter
}
\examples{
 set.seed(1234)
 Gen.Data <- simulate_gendata(n=100, p=200)

}
\author{
Lars Ronnegard
}
