% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothSNPmatrix.R
\name{SmoothSNPmatrix}
\alias{SmoothSNPmatrix}
\title{Imputes column means to missing genotypes}
\usage{
SmoothSNPmatrix(SNP)
}
\arguments{
\item{SNP}{A matrix including SNP coding.}
}
\description{
Imputes column means to missing genotypes.
}
\author{
Lars Ronnegard
}
