% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.data.R
\docType{class}
\name{snp.data}
\alias{snp.data}
\title{An S4 class to represent SNP data}
\description{
An S4 class to represent SNP data
}
\section{Slots}{

\describe{
\item{\code{nids}}{The number of ids as numeric}

\item{\code{idnames}}{The idnames as character}

\item{\code{nsps}}{The number of SNPs as numeric}

\item{\code{snpnames}}{The SNP names as character}

\item{\code{map}}{The order of the SNPS as numeric}

\item{\code{chromosome}}{The chromosome names for each SNP as numeric}

\item{\code{gtps}}{The matrix with SNP coding}
}}

