% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpnames.R
\name{snpnames}
\alias{snpnames}
\title{Extracts the snpnames}
\usage{
snpnames(genabel.data)
}
\arguments{
\item{genabel.data}{A GenABEL-like data of class gwaa.data2.}
}
\description{
Gets the SNP names.
}
\author{
Lars Ronnegard
}
