% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_scan_gwaa2.R
\name{summary.scan.gwaa2}
\alias{summary.scan.gwaa2}
\title{Summary function for the rGLS output}
\usage{
\method{summary}{scan.gwaa2}(object, ...)
}
\arguments{
\item{object}{A scan.gwaa2 object created by the rGLS function}

\item{...}{Possible additional parameters (not used in the current version)}
}
\value{
Returns a data frame with estimated SNP effects, standard errors, test-statistic values, p-values, and corrected p-values.
}
\description{
Creates a Manhattan plot using a slimmed version of the summary.scan.gwaa() function in the GenABEL package
}
\author{
Lars Ronnegard
}
