\name{thinned.flycatchers}
\alias{thinned.flycatchers}
\docType{data}
\title{
10,000 SNPs from 200 collared flycatchers}
\description{
A GenABEL object of class gwaa.data. Only ID and sex included in 
phdata(thinned.flycatchers).
}
\usage{data(thinned.flycatchers)}
\examples{
data(thinned.flycatchers)
summary(thinned.flycatchers)
}
\keyword{datasets}
