% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute.GRM.R
\name{compute.GRM}
\alias{compute.GRM}
\title{Computes a Genetic Relationship Matrix from a GenABEL object}
\usage{
compute.GRM(gen.data, method = "GenABEL")
}
\arguments{
\item{gen.data}{The GenABEL object.}

\item{method}{Method to be used.}
}
\description{
Two alternative methods for GRM computations implemented.
}
\author{
Lars Ronnegard
}

