% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructV.R
\name{constructV}
\alias{constructV}
\title{Constructs the (co)variance matrix for y}
\usage{
constructV(Z, RandC, ratio)
}
\arguments{
\item{Z}{The incidence matrix for the random effects column binded with the Cholesky of the GRM}

\item{RandC}{The number of columns in the two matrices combined in Z.}

\item{ratio}{The ratios between random effect variances and the residual variance.}
}
\description{
Constructs the (co)variance matrix for y.
}
\author{
Lars Ronnegard
}

